requirements.txt
streamlit
python-dotenv
pypdf2
langchain-huggingface
langchain-community
sentence-transformers
faiss-cpu
transformers
